Refactor/server configs (#41)
* chore(tsconfig): rename @configs/* path to @configs

* refactor: move env imports to envVariables file

* refactor: change from export default to name export

* feat: add index.ts to configs folder

* refactor: rename import @configs/[...] to @configs
