+++ server/src/configs/index.ts
@@ -1,0 +1,4 @@
+export * from "./database";
+export * from "./envVariables";
+export * from "./globalPaths";
+export * from "./globalVariables";
