+++ server/src/models/authModel.ts
@@ -1,6 +1,6 @@
 import { Model, model, Schema } from "mongoose";
 import { AuthDocument } from "./types";
-import { encryptionKey } from "@configs/envVariables";
+import { encryptionKey } from "@configs";
 import { encryptToken } from "@utils";
 
 const AuthSchema: Schema<AuthDocument> = new Schema(
