+++ server/tsconfig.json
@@ -4,7 +4,7 @@
       "@libs/*": ["../../libs/*"],
       "@routes/*": ["routes/*"],
       "@controllers/*": ["controllers/*"],
-      "@configs/*": ["configs/*"],
+      "@configs": ["configs"],
       "@middlewares/*": ["middlewares/*"],
       "@models": ["models"],
       "@services": ["services"],
